make channel lazyprop
