+++ rasa/core/domain.py
@@ -5,6 +5,7 @@ import os
 import typing
 from typing import Any, Dict, List, Optional, Text, Tuple, Union, Set
 
+import rasa.utils.common as common_utils
 import rasa.utils.io
 from rasa.cli.utils import bcolors
 from rasa.constants import DOMAIN_SCHEMA_FILE
@@ -299,20 +300,20 @@ class Domain(object):
 
         return int(text_hash, 16)
 
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def user_actions_and_forms(self):
         """Returns combination of user actions and forms"""
 
         return self.user_actions + self.form_names
 
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def num_actions(self):
         """Returns the number of available actions."""
 
         # noinspection PyTypeChecker
         return len(self.action_names)
 
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def num_states(self):
         """Number of used input states for the action prediction."""
 
@@ -387,7 +388,7 @@ class Domain(object):
             return None
 
     # noinspection PyTypeChecker
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def slot_states(self) -> List[Text]:
         """Returns all available slot state strings."""
 
@@ -398,28 +399,28 @@ class Domain(object):
         ]
 
     # noinspection PyTypeChecker
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def prev_action_states(self) -> List[Text]:
         """Returns all available previous action state strings."""
 
         return [PREV_PREFIX + a for a in self.action_names]
 
     # noinspection PyTypeChecker
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def intent_states(self) -> List[Text]:
         """Returns all available previous action state strings."""
 
         return ["intent_{0}".format(i) for i in self.intents]
 
     # noinspection PyTypeChecker
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def entity_states(self) -> List[Text]:
         """Returns all available previous action state strings."""
 
         return ["entity_{0}".format(e) for e in self.entities]
 
     # noinspection PyTypeChecker
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def form_states(self) -> List[Text]:
         return ["active_form_{0}".format(f) for f in self.form_names]
 
@@ -428,12 +429,12 @@ class Domain(object):
 
         return self.input_state_map.get(state_name)
 
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def input_state_map(self) -> Dict[Text, int]:
         """Provides a mapping from state names to indices."""
         return {f: i for i, f in enumerate(self.input_states)}
 
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def input_states(self) -> List[Text]:
         """Returns all available states."""
 
@@ -686,7 +687,7 @@ class Domain(object):
         """Return the configuration for an intent."""
         return self.intent_properties.get(intent_name, {})
 
-    @utils.lazyproperty
+    @common_utils.lazyproperty
     def intents(self):
         return sorted(self.intent_properties.keys())
